back-end/requirements.txt
Flask==2.3.2
yt-dlp==2023.10.6
whisper==20230313
pykakasi==2.1.1
psycopg2-binary==2.9.6
fugashi==1.3.0
requests==2.31.0
meCab-python3